Extract incrementalmerkletree::testing module to a separate `incrementalmerkletree-testing` crate.
