+++ incrementalmerkletree-testing/README.md
@@ -1,0 +1,6 @@
+incrementalmerkletree-testing
+=============================
+
+This crate contains common testing infrastructure used in the implementation of
+the `bridgetree` and `shardtree` crates.
+
